// client/src/clients/deal-system-client-definitions.ts
export class UpdatableListItemDto implements IUpdatableListItemDto {
    deleted!: boolean;
    updated!: boolean;
    id?: number | undefined;

    constructor(data?: IUpdatableListItemDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            deleted: observable,
            updated: observable,
            id: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.deleted = _data["deleted"];
            this.updated = _data["updated"];
            this.id = _data["id"];
        }
    }

    static fromJS(data: any): UpdatableListItemDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdatableListItemDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["deleted"] = this.deleted;
        data["updated"] = this.updated;
        data["id"] = this.id;
        return data; 
    }

    clone(): UpdatableListItemDto {
        const json = this.toJSON();
        let result = new UpdatableListItemDto();
        result.init(json);
        return result;
    }
}

export interface IUpdatableListItemDto {
    deleted: boolean;
    updated: boolean;
    id?: number | undefined;
}
// ...
export class UpdatableOfString implements IUpdatableOfString {
    updated!: boolean;
    value!: string;

    constructor(data?: IUpdatableOfString) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            updated: observable,
            value: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.updated = _data["updated"];
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): UpdatableOfString {
        data = typeof data === 'object' ? data : {};
        let result = new UpdatableOfString();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["updated"] = this.updated;
        data["value"] = this.value;
        return data; 
    }

    clone(): UpdatableOfString {
        const json = this.toJSON();
        let result = new UpdatableOfString();
        result.init(json);
        return result;
    }
}

export interface IUpdatableOfString {
    updated: boolean;
    value: string;
}
// ...
export class UpdatableOfNullableInteger implements IUpdatableOfNullableInteger {
    updated!: boolean;
    value?: number | undefined;

    constructor(data?: IUpdatableOfNullableInteger) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            updated: observable,
            value: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.updated = _data["updated"];
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): UpdatableOfNullableInteger {
        data = typeof data === 'object' ? data : {};
        let result = new UpdatableOfNullableInteger();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["updated"] = this.updated;
        data["value"] = this.value;
        return data; 
    }

    clone(): UpdatableOfNullableInteger {
        const json = this.toJSON();
        let result = new UpdatableOfNullableInteger();
        result.init(json);
        return result;
    }
}

export interface IUpdatableOfNullableInteger {
    updated: boolean;
    value?: number | undefined;
}
// ...
export interface ILookupRequest {
    id: number;
    name: string;
    description: string;
    active?: boolean | undefined;
}
// ...
export class DealAttachmentDto extends UpdatableListItemDto implements IDealAttachmentDto {
    attachmentTypeId!: UpdatableOfNullableInteger;
    attachmentTypeOtherText!: UpdatableOfString;
    linkType?: DealAttachmentLinkType | undefined;
    versions!: DealAttachmentVersionDto[];

    constructor(data?: IDealAttachmentDto) {
        super(data);
        if (data) {
            this.attachmentTypeId = data.attachmentTypeId && !(<any>data.attachmentTypeId).toJSON ? new UpdatableOfNullableInteger(data.attachmentTypeId) : <UpdatableOfNullableInteger>this.attachmentTypeId; 
            this.attachmentTypeOtherText = data.attachmentTypeOtherText && !(<any>data.attachmentTypeOtherText).toJSON ? new UpdatableOfString(data.attachmentTypeOtherText) : <UpdatableOfString>this.attachmentTypeOtherText; 
        }
        if (!data) {
            this.attachmentTypeId = new UpdatableOfNullableInteger();
            this.attachmentTypeOtherText = new UpdatableOfString();
            this.versions = [];
        }
        makeObservable(this, {
            attachmentTypeId: observable,
            attachmentTypeOtherText: observable,
            linkType: observable,
            versions: observable,
        })
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.attachmentTypeId = _data["attachmentTypeId"] ? UpdatableOfNullableInteger.fromJS(_data["attachmentTypeId"]) : new UpdatableOfNullableInteger();
            this.attachmentTypeOtherText = _data["attachmentTypeOtherText"] ? UpdatableOfString.fromJS(_data["attachmentTypeOtherText"]) : new UpdatableOfString();
            this.linkType = _data["linkType"];
            if (Array.isArray(_data["versions"])) {
                this.versions = [] as any;
                for (let item of _data["versions"])
                    this.versions!.push(DealAttachmentVersionDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): DealAttachmentDto {
        data = typeof data === 'object' ? data : {};
        let result = new DealAttachmentDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["attachmentTypeId"] = this.attachmentTypeId ? this.attachmentTypeId.toJSON() : <any>undefined;
        data["attachmentTypeOtherText"] = this.attachmentTypeOtherText ? this.attachmentTypeOtherText.toJSON() : <any>undefined;
        data["linkType"] = this.linkType;
        if (Array.isArray(this.versions)) {
            data["versions"] = [];
            for (let item of this.versions)
                data["versions"].push(item.toJSON());
        }
        super.toJSON(data);
        return data; 
    }

    clone(): DealAttachmentDto {
        const json = this.toJSON();
        let result = new DealAttachmentDto();
        result.init(json);
        return result;
    }
}

export interface IDealAttachmentDto extends IUpdatableListItemDto {
    attachmentTypeId: IUpdatableOfNullableInteger;
    attachmentTypeOtherText: IUpdatableOfString;
    linkType?: DealAttachmentLinkType | undefined;
    versions: DealAttachmentVersionDto[];
}

export enum DealAttachmentLinkType {
    DelegatedTradingAuthorityEvidence = 1,
}

export class DealAttachmentVersionDto extends UpdatableListItemDto implements IDealAttachmentVersionDto {
    fileName!: string;
    fileExtension!: string;
    fileBase64!: string;
    fileSizeInBytes!: number;
    createdDate!: moment.Moment;
    creationUserId!: number;
    creationUserName!: string;
    isLocked!: boolean;

    constructor(data?: IDealAttachmentVersionDto) {
        super(data);
        makeObservable(this, {
            fileName: observable,
            fileExtension: observable,
            fileBase64: observable,
            fileSizeInBytes: observable,
            createdDate: observable,
            creationUserId: observable,
            creationUserName: observable,
            isLocked: observable,
        })
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.fileName = _data["fileName"];
            this.fileExtension = _data["fileExtension"];
            this.fileBase64 = _data["fileBase64"];
            this.fileSizeInBytes = _data["fileSizeInBytes"];
            this.createdDate = _data["createdDate"] ? moment.parseZone(_data["createdDate"].toString()) : <any>undefined;
            this.creationUserId = _data["creationUserId"];
            this.creationUserName = _data["creationUserName"];
            this.isLocked = _data["isLocked"];
        }
    }

    static fromJS(data: any): DealAttachmentVersionDto {
        data = typeof data === 'object' ? data : {};
        let result = new DealAttachmentVersionDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["fileName"] = this.fileName;
        data["fileExtension"] = this.fileExtension;
        data["fileBase64"] = this.fileBase64;
        data["fileSizeInBytes"] = this.fileSizeInBytes;
        data["createdDate"] = this.createdDate ? this.createdDate.toISOString(true) : <any>undefined;
        data["creationUserId"] = this.creationUserId;
        data["creationUserName"] = this.creationUserName;
        data["isLocked"] = this.isLocked;
        super.toJSON(data);
        return data; 
    }

    clone(): DealAttachmentVersionDto {
        const json = this.toJSON();
        let result = new DealAttachmentVersionDto();
        result.init(json);
        return result;
    }
}

export interface IDealAttachmentVersionDto extends IUpdatableListItemDto {
    fileName: string;
    fileExtension: string;
    fileBase64: string;
    fileSizeInBytes: number;
    createdDate: moment.Moment;
    creationUserId: number;
    creationUserName: string;
    isLocked: boolean;
}
// ...
export class AttachmentTypeLookupRequest extends LookupRequest implements IAttachmentTypeLookupRequest {
    other!: boolean;

    constructor(data?: IAttachmentTypeLookupRequest) {
        super(data);
        makeObservable(this, {
            other: observable,
        })
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.other = _data["other"];
        }
    }

    static fromJS(data: any): AttachmentTypeLookupRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AttachmentTypeLookupRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["other"] = this.other;
        super.toJSON(data);
        return data; 
    }

    clone(): AttachmentTypeLookupRequest {
        const json = this.toJSON();
        let result = new AttachmentTypeLookupRequest();
        result.init(json);
        return result;
    }
}

export interface IAttachmentTypeLookupRequest extends ILookupRequest {
    other: boolean;
}